requirements.txt
aiohttp==1.0.5
astroid==2.0.2
async-timeout==2.0.0
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
Cython==0.28.5
isort==4.2.15
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==2.2.2
mccabe==0.6.1
multidict==3.3.2
numpy==1.15.0
opencv-contrib-python==3.3.0.10
opencv-python==3.4.2.17
pandas==0.23.4
Pillow==5.2.0
pylint==2.1.1
pyparsing==2.2.0
pytesseract==0.2.4
python-dateutil==2.6.1
pytz==2018.3
scikit-learn==0.19.2
six==1.11.0
tesseract==0.1.3
typed-ast==1.1.0
typing==3.6.4
virtualenv==15.1.0
websockets==3.4
wrapt==1.10.11
